requirements.txt
attr==0.3.2
beartype==0.12.0
beautifulsoup4==4.13.4
browsergym==0.13.3
GitPython==3.1.44
gradio==5.29.0
gymnasium==1.1.1
huggingface_hub==0.30.2
hydra-core==1.3.2
ipython==8.12.3
joblib==1.4.2
langchain==0.3.25
langchain_community==0.3.23
libwebarena==0.0.4
matplotlib==3.10.1
networkx==3.4.2
numpy==2.2.5
omegaconf==2.3.0
openai==1.77.0
pandas==2.2.3
Pillow==11.2.1
playwright==1.51.0
pydantic==2.11.4
python-dotenv==1.1.0
python-slugify==8.0.4
PyYAML==6.0.2
ray==2.45.0
Requests==2.32.3
setuptools==80.0.0
tapeagents==0.1.10
tiktoken==0.9.0
tqdm==4.67.1
transformers==4.51.3
tweepy==4.15.0
